moved all patient mapping into it own class
